Fiddled with the operand modes for review
We will have 32 modes. That is a lot, but allows to use more than 4 index registers, which I would like to have. Also, the register fields are now 4 bits wide, although we only use 3 bits for now. But if we ever decide to feature 16 general regs, the instruction layout does not need to change again.