Document and Pipeline work
Cleaned up the pipeline stages code. Next stop TLB and MEM objects.

In the document I added a comment on the BRK instruction. How about we just treat BRL 0,0 as a NOP. While this helps with NOPs in the pipeline, it may mean that we do not detect zero instructions words. TDB...