Final update to Document before merge back to main.
Fiddled with MR instruction. It was confusing with the option flags. MR is no MFSR, MTSR, MFCR and MTCR. The instruction stayed the same, but the bits are set accordingly. This model could also be used for some other instructions so that we have less ".<opt>" notations, while still keeping the underlying HW decode. It should also make the assembler decode process easier...