First cut at EX stage
Started to adapt to new instructions formats. Also, cleaned it up. Each instruction is directly executed in the case branch using the pipeline registers directly. This can for sure be done better from a performance point of view, but I want to stay as close as possible to the pipeline work. Otherwise, we could have skipped the whole pipeline mimicking and just execute an instruction.