Some fiddlig with command window output buffer displays
We add the lines to output buffer. So far so good. But how do we treat cursor up and down properly ? They are not really commands and should somehow find the way to the command window with prompt and evaluation... to think about ...

Also, there are still bugs in command history and redo editing ....